Revert "refactor: convert views.py into a module"
This reverts commit d23343896ccbb372fc9e4ed5ab66a719070b4fdd.
